refactor(jssdk): better naming & file structure
