+++ frontend/packages/sdk/src/index.ts
@@ -10,7 +10,8 @@ export * from './contracts/PinkLoggerContract'
 export * from './OnChainRegistry'
 export * from './options'
 export * from './abis/fetchers'
-export * from './pruntime/eip712'
+export * from './utils/eip712'
+export * from './utils/addressConverter'
 export * from './providers/types'
 export * from './providers/EvmAccountMappingProvider'
 export * from './providers/UIKeyringProvider'
